farm_management/requirements.txt
asgiref==3.2.8
attrs==19.3.0
cffi==1.14.0
coverage==5.1
cryptography==2.9.2
Django==3.0.7
django-rest-knox==4.1.0
djangorestframework==3.11.0
importlib-metadata==1.6.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.2
pycparser==2.20
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
wcwidth==0.2.4
zipp==3.1.0
